Update part1b.md
"When calling the method through a reference the value of this becomes the so-called global object and the end result is often not what the software developer had originally intended."

Hello, this is not completely true. If we have two objects and one has a property that references the method of the other, when we invoke the method from the object with the reference, the `this` context becomes the object itself and not the global object.

Thanks for the great curriculum btw!